refactor: separate license checking
