+++ .github/workflows/ci.yml
@@ -110,7 +110,9 @@ jobs:
 
       - name: lint.js
         if: matrix.kind == 'lint'
-        run: deno run --unstable --allow-write --allow-read --allow-run ./tools/lint.js
+        run: |
+          deno run --unstable --allow-write --allow-read --allow-run ./tools/lint.js
+          deno run --unstable --allow-read --allow-run ./tools/check_license.js
 
       - name: test_format.js
         if: matrix.kind == 'lint'
